id is plain integer in Strava -> changed it to Int from Int64 in ios client

+++ Strava/Sources/StravaObject.swift
@@ -10,7 +10,7 @@ import AlamofireObjectMapper
 import ObjectMapper
 
 public class StravaObject: Mappable {
-    var id: Int64?
+    var id: Int?
     var resourceState: Int?
     
     required public init?(map: Map){}
